# docs-site/openapi.yaml
openapi: 3.0.3
info:
  title: LogPulse API
  version: 1.1.0
  description: |
    REST API for LogPulse — a centralized log monitoring platform.

    Three authentication realms exist:

    | Realm | Mechanism | Used by |
    |-------|-----------|---------|
    | Dashboard JWT | `Authorization: Bearer <jwt>` (24h expiry, HS256) | everything under `/api/**` except ingest |
    | Ingest API key | `Authorization: Bearer <application API key>` | `POST /api/logs` only |
    | OAuth redirect | browser 302 flow | `/api/auth/{google,github}/*` |

    There is no public signup: an admin invites users, and the invited person
    activates the account via `POST /api/auth/accept-invite` or by signing in
    with Google/GitHub.

    Errors are always `{"error": "<human readable message>"}`.
servers:
  - url: http://localhost:8080
    description: Local development
  # INFERRED: production URL depends on your deployment; set your own.
  - url: https://your-logpulse-host
    description: Production (replace with your host)
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Email/password sign-in
      description: >
        Issues a JWT valid for 24 hours. Deactivated accounts get 403;
        a pending invite cannot sign in until accepted. Unknown emails and
        wrong passwords return the same 401 message on purpose.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Session created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Invalid credentials
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '403':
          description: Invite not accepted yet, or account deactivated
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /api/auth/accept-invite:
    post:
      tags: [Auth]
      summary: Activate an invited account
      description: >
        Consumes the one-time invite token (valid 7 days), sets name and
        password, flips the account to active and returns a JWT.
      operationId: acceptInvite
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, name, password]
              properties:
                token: { type: string, description: Raw token from the invite link }
                name: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Account activated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '400':
          description: Invalid/used/expired invite link, or validation error
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      operationId: getMe
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: The authenticated caller
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback
      description: >
        Validates the CSRF state cookie, exchanges the code, then redirects
        (302) to `{FRONTEND_URL}/oauth-callback?token=<jwt>`. Existing users
        only — an unknown email is sent back to the login page with
        `?error=...`. A pending invite is satisfied by OAuth sign-in.
      operationId: googleCallback
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        '302':
          description: Redirect to FRONTEND_URL with token or error query param
...
  /api/auth/github/callback:
    get:
      tags: [Auth]
      summary: GitHub OAuth callback
      description: Same behaviour as the Google callback.
      operationId: githubCallback
      security: []
      parameters:
        - { name: code, in: query, schema: { type: string } }
        - { name: state, in: query, schema: { type: string } }
      responses:
        '302':
          description: Redirect to FRONTEND_URL with token or error query param

  # --------------------------------------------------------------- Logs ----
  /api/logs:
    post:
      tags: [Logs]
      summary: Ingest a log entry
      description: >
        Authenticated by the **application API key** (not a user JWT). The
        message is fingerprinted at ingest (variables normalized) so the
        clustering endpoints can group near-duplicates. Levels are
        normalized: `warn`→`warning`, `err`→`error`, `info`→`info`,
        `crit`→`critical`; any other value is stored as-is.
      operationId: ingestLog
      security: [{ ingestApiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_name, level, message]
              properties:
                app_name: { type: string, example: payments-api }
                level:
                  type: string
                  description: 'info | warning | error | critical | debug (normalized)'
                  example: error
                message: { type: string }
                meta:
                  type: object
                  additionalProperties: true
                  description: Free-form structured fields; become queryable evidence.
      responses:
        '201':
          description: Stored (and broadcast over /ws/logs)
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, example: stored }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401':
          description: Missing or invalid API key
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '500':
          description: MongoDB write failed
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      tags: [Logs]
      summary: Search & filter logs
      description: >
        Returns the newest 200 matching entries (timestamp descending). All
        filters are optional and combined with AND.
      operationId: listLogs
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: app, in: query, schema: { type: string }, description: Exact app_name match }
        - { name: level, in: query, schema: { type: string }, description: Exact level match }
        - { name: q, in: query, schema: { type: string }, description: Case-insensitive substring match on message }
      responses:
        '200':
          description: Matching log entries
          content:
            application/json:
              schema:
                type: array
                maxItems: 200
                items: { $ref: '#/components/schemas/LogEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # -------------------------------------------------- Stats & Monitoring ----
  /api/stats/overview:
    get:
      tags: [Stats & Monitoring]
      summary: Dashboard headline stats
      description: 24h log/error totals, error-rate delta vs the previous day, active app count, plus Postgres counters (open incidents, suppressed alerts in the last hour).
      operationId: getOverview
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Overview snapshot
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StatsOverview' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/stats/timeseries:
    get:
      tags: [Stats & Monitoring]
      summary: Volume timeseries
      description: >
        Fixed-width zero-filled buckets. Bucketing is computed in Go from
        per-minute Mongo groups, so it works on any MongoDB version.
      operationId: getTimeseries
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: app, in: query, schema: { type: string } }
        - { name: hours, in: query, schema: { type: integer, default: 24, maximum: 720 }, description: Lookback window }
        - name: bucket
          in: query
          schema: { type: string, enum: [minute, hour, day], default: hour }
      responses:
        '200':
          description: Chronological buckets
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TimeseriesPoint' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/stats/top-apps:
    get:
      tags: [Stats & Monitoring]
      summary: Busiest applications
      operationId: getTopApps
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: hours, in: query, schema: { type: integer, default: 24 } }
      responses:
        '200':
          description: Up to 8 apps, total volume descending
          content:
            application/json:
              schema:
                type: array
                maxItems: 8
                items: { $ref: '#/components/schemas/TopApp' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/stats/levels:
    get:
      tags: [Stats & Monitoring]
      summary: Level counts over a window
      operationId: getLevelCounts
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: hours, in: query, schema: { type: integer, default: 24 } }
        - { name: app, in: query, schema: { type: string } }
      responses:
        '200':
          description: 'Map of level → count, e.g. {"error": 607, "info": 7488}'
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/clusters:
    get:
      tags: [Stats & Monitoring]
      summary: Failure-signature clusters (AI-1)
      description: >
        Groups log lines by their normalized pattern fingerprint. Only
        entries ingested after fingerprinting shipped have a pattern_hash.
        Sorted worst-level first, then by count.
      operationId: getClusters
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: hours, in: query, schema: { type: integer, default: 24 } }
        - { name: app, in: query, schema: { type: string } }
        - { name: level, in: query, schema: { type: string } }
        - { name: min_count, in: query, schema: { type: integer, default: 2, minimum: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
      responses:
        '200':
          description: Clusters, largest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Cluster' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/anomalies:
    get:
      tags: [Stats & Monitoring]
      summary: Live anomaly sweep (AI-2)
      description: >
        Computes anomalies on request: the last complete hour vs the same
        hour-of-day across the past week (volume z-score, error-rate Wald
        z-test), plus a raw 15-minute silence check. The background detector
        persists findings as incidents; this endpoint is read-only.
      operationId: getAnomalies
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current findings (may be empty)
          content:
            application/json:
              schema:
                type: object
                required: [results, note]
                properties:
                  results:
                    type: array
                    items: { $ref: '#/components/schemas/AnomalyResult' }
                  note: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/forecast:
    get:
      tags: [Stats & Monitoring]
      summary: Volume forecast, next 24h (AI-7)
      description: >
        Linear trend + hour-of-day profile over the past week; the band is
        ±1.5σ of fit residuals. Needs at least 24h of history — otherwise
        `points` is empty and `note` explains why.
      operationId: getForecast
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: app, in: query, schema: { type: string }, description: Omit for all apps }
      responses:
        '200':
          description: Forecast with trailing history
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Forecast' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------------ Incidents ----
  /api/incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      description: Newest first, max 200. Without a status filter, all statuses are returned.
      operationId: listIncidents
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, open, acknowledged, resolved] }
          description: '`active` = open + acknowledged.'
      responses:
        '200':
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Incident' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Incidents]
      summary: Declare an incident manually
      operationId: createIncident
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string }
                severity: { type: string, enum: [critical, error, warning, info], default: error }
                affected_apps:
                  type: string
                  description: JSON-encoded array of app names, e.g. '["api","worker"]'
                summary: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/incidents/{id}:
    get:
      tags: [Incidents]
      summary: Get one incident
      operationId: getIncident
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IncidentId' }]
      responses:
        '200':
          description: The incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Incidents]
      summary: Update status / severity / fields
      description: >
        Setting status to `resolved` stamps `resolved_at`; moving back to
        open/acknowledged clears it.
      operationId: updateIncident
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IncidentId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [open, acknowledged, resolved] }
                severity: { type: string, enum: [critical, error, warning, info] }
                title: { type: string }
                summary: { type: string }
      responses:
        '200':
          description: Updated incident
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Incident' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/incidents/{id}/analysis:
    get:
      tags: [Incidents]
      summary: Root-cause analysis (AI-5)
      description: >
        Heuristic, evidence-first analysis computed on request: 5-minute
        timeline, dominant failure signatures, structured-field hints, and
        ranked hypotheses. Every hypothesis cites the log lines it is based
        on and is a suggestion, never an auto-action.
      operationId: getIncidentAnalysis
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/IncidentId' }]
      responses:
        '200':
          description: Analysis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CopilotAnalysis' }
        '404': { $ref: '#/components/responses/NotFound' }

  # --------------------------------------------------------------- Alerts ----
  /api/alerts:
    get:
      tags: [Alerts]
      summary: List alert rules
      operationId: listAlerts
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: application_id, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Rules with the resolved application name
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AlertRuleView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Alerts]
      summary: Create an alert rule
      operationId: createAlert
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [application_id, level, threshold, window_seconds]
              properties:
                application_id: { type: integer }
                name: { type: string, description: 'Defaults to "<level> threshold" when omitted' }
                level: { type: string, example: error }
                keyword: { type: string, description: Optional substring match }
                threshold: { type: integer, minimum: 1 }
                window_seconds: { type: integer, minimum: 1 }
                cooldown_seconds:
                  type: integer
                  description: 'Repeat breaches inside the cooldown are recorded but suppressed (AI-6). Defaults to 300.'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertRule' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/alerts/parse:
    post:
      tags: [Alerts]
      summary: Parse plain language into a draft rule (AI-8)
      description: >
        Deterministic parser — no LLM, no data leaves the deployment. Returns
        a draft for human review; nothing is activated until you POST it to
        `/api/alerts`.
      operationId: parseAlert
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: Alert me when payments-api has more than 5 errors mentioning timeout within 2 minutes
      responses:
        '200':
          description: Draft rule plus the matched application id (or null)
          content:
            application/json:
              schema:
                type: object
                properties:
                  draft: { $ref: '#/components/schemas/ParsedAlertDraft' }
                  application_id:
                    type: integer
                    nullable: true
        '400': { $ref: '#/components/responses/ValidationError' }
  /api/alerts/{id}:
    patch:
      tags: [Alerts]
      summary: Update a rule (partial)
      operationId: updateAlert
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/AlertId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                level: { type: string }
                keyword: { type: string }
                threshold: { type: integer }
                window_seconds: { type: integer }
                cooldown_seconds: { type: integer }
                enabled: { type: boolean }
      responses:
        '200':
          description: Updated rule
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlertRule' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Alerts]
      summary: Delete a rule
      operationId: deleteAlert
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/AlertId' }]
      responses:
        '200': { $ref: '#/components/responses/Deleted' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/alerts/events:
    get:
      tags: [Alerts]
      summary: Alert trigger history
      description: Newest first, max 100. Includes suppressed (triaged) events.
      operationId: listAlertEvents
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: rule_id, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AlertEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # --------------------------------------------------------- Applications ----
  /api/applications:
    get:
      tags: [Applications]
      summary: List registered applications (with API keys)
      operationId: listApplications
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Applications, newest first
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Application' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Applications]
      summary: Register an application
      description: Generates a fresh API key (crypto/rand, 64 hex chars).
      operationId: createApplication
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Created — store the api_key now; it is shown in the list later too
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Application' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------- Saved Searches ----
  /api/saved-searches:
    get:
      tags: [Saved Searches]
      summary: List the caller's saved searches
      operationId: listSavedSearches
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Searches, updated_at descending
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SavedSearch' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Saved Searches]
      summary: Save a filter combination
      description: '`filters` is a JSON blob of query params (e.g. `{"level":"error","q":"timeout"}`).'
      operationId: createSavedSearch
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, filters]
              properties:
                name: { type: string }
                filters: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedSearch' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/dashboards:
    get:
      tags: [Dashboards]
      summary: List the caller's dashboards
      operationId: listDashboards
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Dashboards, updated_at descending
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Dashboard' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Dashboards]
      summary: Create a dashboard
      operationId: createDashboard
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Dashboard' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/reports/{id}:
    delete:
      tags: [Reports]
      summary: Delete a report (owner only)
      operationId: deleteReport
      security: [{ bearerAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/ReportId' }]
      responses:
        '200': { $ref: '#/components/responses/Deleted' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Dashboard JWT from /api/auth/login — 24h expiry (HS256).
    ingestApiKey:
      type: http
      scheme: bearer
      description: Per-application API key from POST /api/applications.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    LoginResponse:
      type: object
      properties:
        token: { type: string, description: JWT, expires in 24h }
        user: { $ref: '#/components/schemas/User' }
...